module8-task1 Оптимизируй это (часть 1) (Правки №1)
