+++ src/components/bookmark-button-component.tsx
@@ -1,25 +1,44 @@
 import {JSX, memo} from 'react';
 import clsx from 'clsx';
-import {BookmarkButtonClasses} from '@/constants/constants.ts';
+import {AppRoute, AuthorizationStatus, BookmarkButtonClasses} from '@/constants/constants.ts';
+import {useAppDispatch, useAppSelector} from '@/hooks';
+import {useNavigate} from 'react-router-dom';
+import {getAuthorizationStatus} from '@/store/user-process/user-process.selectors.ts';
+import {fetchFavoritesStatusAction} from '@/store/api-actions.ts';
+import {getFavorites} from '@/store/main-data/main-data.selectors.ts';
 
 
 type BookmarkButtonProps = {
   width: string;
   height: string;
-  isFavorite: boolean;
   classType: 'card' | 'offer';
-  handleClick: () => void;
+  offerId: string;
 }
 
 function BookmarkButtonComponent({
   width,
   height,
-  isFavorite,
   classType,
-  handleClick
+  offerId,
 }: BookmarkButtonProps): JSX.Element {
+
+  const favorites = useAppSelector(getFavorites);
+  const isFavorite = favorites.map(({id}) => id).includes(offerId);
+
   const buttonText = isFavorite ? 'In bookmarks' : 'To bookmarks';
   const buttonClasses = BookmarkButtonClasses[classType];
+
+  const dispatch = useAppDispatch();
+  const navigate = useNavigate();
+  const authorizationStatus = useAppSelector(getAuthorizationStatus);
+
+  const handleClick = () => {
+    if (authorizationStatus !== AuthorizationStatus.Auth) {
+      navigate(AppRoute.Login);
+      return;
+    }
+    dispatch(fetchFavoritesStatusAction({offerId: offerId, isFavorite: !isFavorite}));
+  };
 
   return (
     <button
