+++ src/pages/offer/offer-page.tsx
@@ -1,14 +1,9 @@
 import {JSX, memo, useEffect, useState} from 'react';
-import {useNavigate, useParams} from 'react-router-dom';
+import {useParams} from 'react-router-dom';
 import PageNotFound from '../error/page-not-found.tsx';
 import {useAppDispatch, useAppSelector} from '@/hooks';
 import {Helmet} from 'react-helmet-async';
-import {
-  fetchCommentsAction,
-  fetchFavoritesStatusAction,
-  fetchNearbyOffersAction,
-  fetchOfferAction
-} from '@/store/api-actions.ts';
+import {fetchCommentsAction, fetchNearbyOffersAction, fetchOfferAction} from '@/store/api-actions.ts';
 import SpinnerComponent from 'components/spinner/spinner-component.tsx';
 import OfferGallery from '@/pages/offer/components/offer-gallery-component.tsx';
 import OfferUserStatusComponent from '@/pages/offer/components/offer-user-status-component.tsx';
@@ -21,7 +16,7 @@ import BadgeOfferMark from 'components/badge-offer-mark-component.tsx';
 import LocationMap from 'components/location-map-component.tsx';
 import OfferNearPlaces from '@/pages/offer/components/offer-near-places-component.tsx';
 import OfferReviewsList from '@/pages/offer/components/offer-reviews-list-component.tsx';
-import {AppRoute, AuthorizationStatus} from '@/constants/constants.ts';
+import {AuthorizationStatus} from '@/constants/constants.ts';
 import OfferReviewsFormComponent from '@/pages/offer/components/offer-reviews-form-component.tsx';
 import {calculateRatingInPercent} from '@/utils/calculation-helper.ts';
 import {getAuthorizationStatus} from '@/store/user-process/user-process.selectors.ts';
@@ -43,7 +38,6 @@ const NEAREST_OFFERS_COUNT = 3;
 
 function OfferPage(): JSX.Element {
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
   const authorizationStatus = useAppSelector(getAuthorizationStatus);
   const isDataLoading = useAppSelector(getOfferLoadingStatus);
   const isCommentsLoading = useAppSelector(getCommentsLoadingStatus);
@@ -66,17 +60,6 @@ function OfferPage(): JSX.Element {
   const nearestFirstThreeOffers = nearestOffers.slice(DEFAULT_START_INDEX, NEAREST_OFFERS_COUNT);
   const comments = useAppSelector(getComments);
 
-  const handleClick = () => {
-    if (!currentOffer) {
-      return;
-    }
-    if (authorizationStatus !== AuthorizationStatus.Auth) {
-      navigate(AppRoute.Login);
-      return;
-    }
-    dispatch(fetchFavoritesStatusAction({offerId: currentOffer.id, isFavorite: !(currentOffer.isFavorite)}));
-  };
-
   if (isInitialLoad || isDataLoading || isNearestLoading) {
     return <SpinnerComponent/>;
   }
@@ -94,7 +77,6 @@ function OfferPage(): JSX.Element {
     price,
     title,
     isPremium,
-    isFavorite
   } = currentOffer;
 
   const {
@@ -120,9 +102,8 @@ function OfferPage(): JSX.Element {
               <BookmarkButton
                 width="31"
                 height="33"
-                isFavorite={isFavorite}
                 classType={'offer'}
-                handleClick={handleClick}
+                offerId={currentOffer.id}
               />
             </div>
             <div className="offer__rating rating">
